requirements.txt
numpy 
pandas
scikit-learn
matplotlib
torch
python-dotenv
langchain
langchain_openai
langchain_community
pypdf 
faiss-cpu
langchain_groq
pytrends
langgraph
pytube
pinecone
langchain-pinecone
streamlit
youtube_transcript_api
openai-whisper
yt-dlp
langchain-chroma